fix(ai-service): mypy --strict 全库清零(1159->0 errors/226 files)
批次 5-9 累计修复:
- app/sio/ untyped-decorator + float(7 errors)
- app/skills/content_engine/ 直接子文件(192 errors)
- app/skills/content_engine/lib/(147 errors)
- app/skills/koubo_workflow/(252 errors)
- app/services/ middleware + core + langgraph(部分)
- app/routers/ + app/api/ + app/middleware/(38 errors)
- app/core/ compaction + llm_gateway(7 errors)

验证: mypy --strict 0 errors/226 files + mypy 非 strict 0 errors/226 files

+++ apps/ai-service/app/services/memory_decay.py
@@ -135,7 +135,7 @@ class MemoryDecayManager:
             days = (now - last).total_seconds() / 86400.0
             if days <= 0:
                 return 1.0
-            return 0.5 ** (days / half_life_days)
+            return float(0.5 ** (days / half_life_days))
         except Exception:
             return 1.0
